refactor(onboarding): rename wallet components to follow step naming convention
